+++ src/components/pages/index.ts
@@ -1,2 +1,3 @@
 export * from './TVDashboard/TVDashboard';
 export * from './HomePage';
+export * from './DevCornerTwoDashboard';