+++ src/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './routes';
 export * from './DASHBOARD_LIST';
+export * from './JIRA_NOVA';