+++ src/types/jira/index.ts
@@ -1,2 +1,3 @@
 export * from './JiraIssue';
 export * from './JiraSearch';
+export * from './NovaAnalytics';